requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crawl4ai==0.4.248
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.6.7
deepdiff==6.7.1
distro==1.9.0
duckduckgo-search==7.5.1
fake-http-header==0.3.5
fake-useragent==2.1.0
filelock==3.18.0
flask==3.1.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.50
langchain-openai==0.3.7
langchain-text-splitters==0.3.7
langgraph==0.2.71
langgraph-checkpoint==2.0.24
langgraph-sdk==0.1.61
langsmith==0.1.147
litellm==1.65.1
lxml==5.3.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mesop==1.0.1
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
openai==1.70.0
ordered-set==4.1.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pillow==10.4.0
playwright==1.51.0
primp==0.14.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pyee==12.1.1
pygments==2.19.1
pyopenssl==25.0.0
python-dotenv==1.1.0
pyyaml==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
sqlalchemy==2.0.35
tenacity==9.1.2
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
urllib3==2.3.0
watchdog==6.0.0
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
